Streaming stability + organization improvements
* Split HLS endpoints to a new file
* Tweak some of the settings for HLS concat playback
* Some initial MPEG-DASH work - this is a WIP
* Tweaks to the Video / useHls frontend stuff to make HLS stop reloading
  so much
