+++ server/src/services/m3uService.ts
@@ -1,6 +1,7 @@
 import { sortBy } from 'lodash-es';
 import { ChannelDB } from '../dao/channelDb.js';
 import { FileCacheService } from './fileCacheService.js';
+import { getSettings } from '../dao/settings.js';
 
 /**
  * Manager and Generate M3U content
@@ -75,6 +76,48 @@ export class M3uService {
     }
 
     return this.replaceHostOnM3u(host, data);
+  }
+
+  async buildChannelM3U(
+    protocol: string,
+    host: string,
+    channel: string | number,
+    sessionId: number,
+  ) {
+    const settings = await getSettings();
+    // Maximum number of streams to concatinate beyond channel starting
+    // If someone passes this number then they probably watch too much television
+    const maxStreamsToPlayInARow = 100;
+
+    const lines: string[] = [
+      '#EXTM3U',
+      '#EXT-X-VERSION:3',
+      '#EXT-X-MEDIA-SEQUENCE:0',
+      '#EXT-X-ALLOW-CACHE:YES',
+      '#EXT-X-TARGETDURATION:60',
+      '#EXT-X-PLAYLIST-TYPE:VOD',
+      // `#EXT-X-STREAM-INF:BANDWIDTH=1123000`,
+    ];
+
+    const ffmpegSettings = settings.ffmpegSettings();
+
+    if (ffmpegSettings.enableTranscoding) {
+      lines.push(
+        `${protocol}://${host}/stream?channel=${channel}&first=0&m3u8=1&session=${sessionId}`,
+      );
+    }
+
+    lines.push(
+      `${protocol}://${host}/stream?channel=${channel}&first=1&m3u8=1&session=${sessionId}`,
+    );
+
+    for (let i = 0; i < maxStreamsToPlayInARow - 1; i++) {
+      lines.push(
+        `${protocol}://${host}/stream?channel=${channel}&m3u8=1&session=${sessionId}`,
+      );
+    }
+
+    return lines.join('\n');
   }
 
   /**
